Clean up Policies

+++ app/Http/Controllers/ServiceTemplateController.php
@@ -232,9 +232,8 @@ class ServiceTemplateController extends Controller
      * Apply specified Service Template to Device Groups.
      *
      * @param  ServiceTemplate  $template
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDeviceGroups(ServiceTemplate $template)
+    private function applyDeviceGroups(ServiceTemplate $template): void
     {
         foreach (DeviceGroup::inServiceTemplate($template->id)->get() as $device_group) {
             foreach (Device::inDeviceGroup($device_group->id)->get() as $device) {
@@ -255,18 +254,14 @@ class ServiceTemplateController extends Controller
                 );
             }
         }
-        $msg = __('Services for Template :name have been updated', ['name' => $template->name]);
-
-        return response($msg, 200);
     }
 
     /**
      * Apply specified Service Template to Devices.
      *
      * @param  ServiceTemplate  $template
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDevices(ServiceTemplate $template)
+    private function applyDevices(ServiceTemplate $template): void
     {
         foreach (Device::inServiceTemplate($template->id)->get() as $device) {
             $device->services()->updateOrCreate(
@@ -285,9 +280,6 @@ class ServiceTemplateController extends Controller
                 ]
             );
         }
-        $msg = __('Services for Template :name have been updated', ['name' => $template->name]);
-
-        return response($msg, 200);
     }
 
     /**
@@ -297,6 +289,8 @@ class ServiceTemplateController extends Controller
      */
     public function applyAll()
     {
+        $this->authorize('update', ServiceTemplate::class);
+
         foreach (ServiceTemplate::all() as $template) {
             $this->apply($template);
         }
@@ -307,17 +301,12 @@ class ServiceTemplateController extends Controller
 
     /**
      * Apply all Service Templates for a device
-     *
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDeviceAll(int $device_id)
+    public function applyDeviceAll(int $device_id): void
     {
         foreach (ServiceTemplate::all() as $template) {
             $this->applyDevice($template, $device_id);
         }
-        $msg = __('All Service Templates have been applied to device ' . $device_id);
-
-        return response($msg, 200);
     }
 
     /**
@@ -328,11 +317,13 @@ class ServiceTemplateController extends Controller
      */
     public function apply(ServiceTemplate $template)
     {
+        $this->authorize('update', ServiceTemplate::class);
+
         if ($template->type == 'dynamic') {
             $template->updateDevices();
         }
-        ServiceTemplateController::applyDevices($template);
-        ServiceTemplateController::applyDeviceGroups($template);
+        $this->applyDevices($template);
+        $this->applyDeviceGroups($template);
 
         // remove any remaining services no longer in the correct device group
         foreach (Device::notInServiceTemplate($template->id)->notInDeviceGroup($template->groups->pluck('id'))->pluck('device_id') as $device_id) {
@@ -348,9 +339,8 @@ class ServiceTemplateController extends Controller
      *
      * @param  ServiceTemplate  $template
      * @param  int  $device_id
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDevice(ServiceTemplate $template, int $device_id)
+    private function applyDevice(ServiceTemplate $template, int $device_id): void
     {
         // Check if the device needs to be added
         foreach (Device::inServiceTemplate($template->id)->where('device_id', $device_id)->get() as $device) {
@@ -370,7 +360,7 @@ class ServiceTemplateController extends Controller
                 ]
             );
 
-            return response('Service template ' . $template->id . ' applied to device ID ' . $device_id, 200);
+            return; // found
         }
 
         foreach (DeviceGroup::inServiceTemplate($template->id)->get() as $device_group) {
@@ -391,7 +381,7 @@ class ServiceTemplateController extends Controller
                     ]
                 );
 
-                return response('Service template ' . $template->id . ' applied to device ID ' . $device_id, 200);
+                return; // found
             }
         }
 
@@ -399,8 +389,6 @@ class ServiceTemplateController extends Controller
         foreach (Device::notInServiceTemplate($template->id)->notInDeviceGroup($template->groups->pluck('id'))->where('device_id', $device_id)->pluck('device_id') as $device_id) {
             Service::where('device_id', $device_id)->where('service_template_id', $template->id)->delete();
         }
-
-        return response('Service template ' . $template->id . ' applied to device ID ' . $device_id, 200);
     }
 
     /**
@@ -411,6 +399,8 @@ class ServiceTemplateController extends Controller
      */
     public function remove(ServiceTemplate $template)
     {
+        $this->authorize('update', ServiceTemplate::class);
+
         Service::where('service_template_id', $template->id)->delete();
 
         $msg = __('All Service Templates have been applied');
